split covid into covid and covidutils
